requirements.txt
requests
markdown
python-dotenv
openai
